Add omitted; make (...) an alias for omitted
'omitted' is a more intuitive name for the universal omitted content
operator.
Experiments show that the frustrated user of 'undefined' is apt to try
'omitted' before trying '(...)'.
The name change has been estimated to increase discoverability by at
least an order of magnitude.

'(...)' is provided to save precious keystrokes for frequent users and
because it looks nicer when printed.