Provide an alternate implementation of undefined
Recent improvements in the theory of the undefinable suggests
an alternative implementation of undefined which more clearly
expresses the idea that something cannot be expressed.

Acme.Omitted.undefined is a sane alternative to Prelude.undefined,
though it relies on the informal contract that the programmer
never use it to stand for the merely omitted.